requirements.txt
asgiref==3.5.2
attrs==21.4.0
backports.zoneinfo==0.2.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
coverage==6.4.1
Django==4.0.5
django-basicauth==0.5.3
django-environ==0.8.1
django-request-logging==0.7.5
idna==3.3
ijson==3.1.4
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytest==7.1.2
requests==2.27.1
requests-mock==1.9.3
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
urllib3==1.26.9
